Sort listed books and allow listing repo-only
