+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__/
 **.DS_Store
 **dist
 sync-osbooks.yml
+.vscode
