+++ pipemgr/src/models/__init__.py
@@ -1,1 +1,4 @@
 from .args import Args
+from .osbook import OSBook
+
+__all__ = [Args, OSBook]
